修复: SpiderWallet联动/论坛树状/群密钥/PIN/多bug修复
- SpiderWallet: 3 PIN系统(解锁/胁迫/反向) + 死人开关(只清数据不发警告) + /api/supported_chains
- SC跳转SW前检查链/币支持, 不支持则提醒不跳转
- Minecraft客户端移除论坛UI(TAB_COUNT 7->6)
- 评论重构为树状本质+扁平热度排序, 显示回复@xxx标识
- 修复自动模式下自动换频率bug(始终使用用户指定频率)
- 修复Android无法输入合法PIN(maxLength 6->16)
- 修复群消息未使用群组密钥加密(实现group_key AES-256-GCM)
- 修复已读回执消息过长不发送(头尾曾经出现过即可)
- 修复加密货币地址含冒号解析错误(反斜杠转义)
- 修复create_user_for_admin用服务器MAC生成UUIDv1(改用uuid4)
- 修复服务端多处导入错误/参数错位/空壳实现/跨服方法缺失
- 全链路测试36+全部通过

+++ android/app/src/main/java/com/spider/android/file/FileTransferManager.kt
@@ -151,7 +151,7 @@ class FileTransferManager(
                 isFile = true,
                 fileName = fileName,
                 filePath = savedPath,
-                fileSize = fileData.size,
+                fileSize = fileData.size.toLong(),
                 mimeType = mimeType,
                 deliveryStatus = "delivered"
             )
